requirements.txt
django
pandas
djangorestframework
markdown
django-filter